+++ src/http/mod.rs
@@ -2,11 +2,11 @@ mod client;
 mod data_loader;
 
 mod data_loader_request;
-mod handler;
 mod http_1;
 mod http_2;
 mod method;
 mod request_context;
+mod request_handler;
 mod response;
 mod server;
 mod server_config;
@@ -18,10 +18,10 @@ use cache_control::CacheControl;
 pub use client::*;
 pub use data_loader::*;
 pub use data_loader_request::*;
-pub use handler::{handle_batch_request, handle_single_request};
 use hyper::header::CACHE_CONTROL;
 pub use method::Method;
 pub use request_context::RequestContext;
+pub use request_handler::{handle_batch_request, handle_single_request};
 pub use response::*;
 pub use server::start_server;
 pub use server_context::ServerContext;
